訂單系統前後台全部更改成功，有部分會員綁定token有先佔存

+++ hotel-backend/src/main/java/com/hotel/model/entity/Product.java
@@ -1,7 +1,5 @@
 package com.hotel.model.entity;
 
-import java.math.BigDecimal;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -31,7 +29,7 @@ public class Product {
     private String description;
 
     @Column(name = "Price")
-    private BigDecimal price;
+    private int price;
 
     @Column(name = "Stock")
     private Integer stock;
@@ -74,11 +72,11 @@ public class Product {
         this.description = description;
     }
 
-    public BigDecimal getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(int price) {
         this.price = price;
     }
